Add camera/time, input enhancements and assets
Introduce Camera and Time classes with camera position/zoom smoothing and delta-time handling; integrate camera and time into Graphics. Extend Input with MouseButton (including scroll) APIs and add mouse state tracking. Expand Raylib texture wrapper and Texture interface (constructors, move semantics, atlas loading, float dimensions, utilities). Update TileMap to support offset and larger tileset/atlas; convert map.tmx to an infinite/chunked layout and update tiles.tsx/tilemap.png. Add numerous PixelArt assets and Unity package. Update main to load the texture atlas and use camera/time for movement and zoom controls. Also minor build tweaks (rebuild script changes and newline fix in CMakeLists).

+++ include/texture.hpp
@@ -12,16 +12,16 @@ namespace EcoSim
         Texture(const std::string filePath) {};
 
         virtual ~Texture() = default;
-        virtual int getWidth() const = 0;
-        virtual int getHeight() const = 0;
+        virtual float getWidth() const = 0;
+        virtual float getHeight() const = 0;
 
         void GetTexture(void *buffer) {};
 
         virtual bool isTextureLoaded() const = 0;
 
     protected:
-        int width;
-        int height;
+        float width;
+        float height;
     };
 }
